# don't allocate beam scratch buffers if in cache