--- a/test/test_search.py
+++ b/test/test_search.py
@@ -13,8 +13,7 @@ class TestTimeLinearizer(unittest.TestCase):
 
   def test_reasonable_time(self):
     si = [si for si in create_schedule([Tensor([1,2,3,4]).add(1).lazydata]) if si.ast.op not in LoadOps][0]
-    rawbufs = [Buffer(Device.DEFAULT, si.out.st.real_size(), si.out.dtype)] + [Buffer(Device.DEFAULT, x.st.real_size(), x.dtype) for x in si.inputs]
-    tm = time_linearizer(Linearizer(si.ast), rawbufs, allow_test_size=False, cnt=10)
+    tm = time_linearizer(Linearizer(si.ast), allow_test_size=False, cnt=10)
     assert tm > 0 and tm != float('inf')
 
 if __name__ == '__main__':