Prevents OOMs when memory is very tight.

Q: If BEAM could finish the first time with scratch buffers, why can't it finish the second time? A: When restoring from resnet checkpoint, all model weights are immediately realized, vs when training from scratch, model weights are loaded and BEAMd one by one.

Another possible solution is to go back to using input buffers in beam, except for the output buffer

The OOM I ran into with resnet is the "soft" kind: #3343 , which happens after just 2 steps! This problem was fixed by #3364 though.

Steps to replicate "soft" OOM (resnet_3 branch):
1. BEAM=2 BS=624 GPUS=6 train resnet, save a checkpoint
2. BEAM=2 RESUME=<ckpt> BS=624 GPUS=6 resume resnet
should crash in 2 steps or so