@geohot 2024-02-15
What happens if the kernel modifies one of it's input buffers.

And why keep the old functionality around? Commit.