@chaosagent 2024-02-21
> What happens if the kernel modifies one of it's input buffers.

this is not poaching live buffers, this just allocates fresh buffers inside beam_search and time_linearizer. LRU should make this fast.