@chaosagent 2024-02-21
> i don't quite understand why this helps OOM. Can you give an example case that it reduces peak memory than before?

my theory is that on fresh init, weights are lazily initialized when the calculation gets to that layer. but on resume, load_state_dict will immediately realize all weights. On fresh init, this means that before all the weights are intialized the beam buffers will not increase the peak global memory, but with restore from checkpoint, they will. This is kind of a tight case to use BEAM with anyways. 

I think we want to eventually implement buffer-poaching, since with the current setup, you need at least your largest buffer's worth of free space to use beam safely. 

with patch and BEAM cached (resnet_3 branch, this is before the leak fix. peak mem is measured by high watermark of GlobalCounters.mem_used):
```
RESUME=./ckpts/20240219_121235_bfn1jsys_e37.safe WANDB_RESUME=bfn1jsys BEAM=2 CKPT_EPOCHS=100 BS=624 GPUS=6 MODEL=resnet WORKERS=32 python examples/mlperf/model_train.py 1000
    0 38719.66 ms run, 38676.02 ms python,   20.60 ms HIP * 6,  23.04 ms fetch data,  1.90 loss, 0.78 acc, 0.078467 LR, 2.55 GB used,    399.69 GFLOPS
step 0 peak global mem: 99.98448     
    1 6768.63 ms run, 6732.26 ms python,    7.08 ms HIP * 6,  29.28 ms fetch data,  1.95 loss, 0.77 acc, 0.078456 LR, 138.02 GB used,   2286.42 GFLOPS
step 1 peak global mem: 128.53910    
    2  502.29 ms run,  321.46 ms python,  157.24 ms HIP * 6,  23.60 ms fetch data,  1.86 loss, 0.77 acc, 0.078445 LR, 138.02 GB used,  30810.56 GFLOPS
step 2 peak global mem: 128.62659    
    3  503.04 ms run,  313.53 ms python,  169.38 ms HIP * 6,  20.13 ms fetch data,  1.86 loss, 0.78 acc, 0.078434 LR, 138.02 GB used,  30764.63 GFLOPS
step 3 peak global mem: 128.62659    
                                                                                                                  
```

without patch:
```
RESUME=./ckpts/20240219_121235_bfn1jsys_e37.safe WANDB_RESUME=bfn1jsys BEAM=2 CKPT_EPOCHS=100 BS=624 GPUS=6 MODEL=resnet WORKERS=32 python examples/mlperf/model_train.py 1000
    0 41277.58 ms run, 41236.19 ms python,   17.54 ms HIP * 6,  23.85 ms fetch data,  1.90 loss, 0.78 acc, 0.078467 LR, 2.55 GB used,    374.92 GFLOPS
step 0 peak global mem: 102.84211    
    1 6700.08 ms run, 6663.82 ms python,    6.86 ms HIP * 6,  29.40 ms fetch data,  1.95 loss, 0.77 acc, 0.078456 LR, 138.02 GB used,   2309.81 GFLOPS
step 1 peak global mem: 128.53910    
    2  505.43 ms run,  331.69 ms python,  152.01 ms HIP * 6,  21.73 ms fetch data,  1.86 loss, 0.78 acc, 0.078445 LR, 138.02 GB used,  30619.32 GFLOPS
step 2 peak global mem: 128.62658    
(before, it would crash here with GPU page fault)
    3  505.47 ms run,  300.50 ms python,  180.12 ms HIP * 6,  24.85 ms fetch data,  1.86 loss, 0.77 acc, 0.078434 LR, 138.02 GB used,  30616.81 GFLOPS
step 3 peak global mem: 128.62659    
    4  498.95 ms run,  302.93 ms python,  174.48 ms HIP * 6,  21.54 ms fetch data,  1.96 loss, 0.77 acc, 0.078423 LR, 138.02 GB used,  31016.98 GFLOPS
step 4 peak global mem: 128.62659    
                                                                                                             
```

without patch, fresh init (with patch, beam uncached, fresh init should behave the same):

```
BEAM=2 CKPT_EPOCHS=100 BS=624 GPUS=6 MODEL=resnet WORKERS=32 python examples/mlperf/model_train.py 1000
    0 40325.78 ms run, 40286.21 ms python,   17.92 ms HIP * 6,  21.66 ms fetch data,  7.04 loss, 0.00 acc, 0.000499 LR, 2.55 GB used,    383.91 GFLOPS
step 0 peak global mem: 102.26374                                                                                                                     
    1 6882.40 ms run, 6846.02 ms python,    6.96 ms HIP * 6,  29.43 ms fetch data,  7.00 loss, 0.00 acc, 0.000748 LR, 138.02 GB used,   2248.62 GFLOPS
step 1 peak global mem: 128.54029                                                                                                                     
    2  529.02 ms run,  299.96 ms python,  177.94 ms HIP * 6,  51.11 ms fetch data,  7.02 loss, 0.00 acc, 0.000997 LR, 138.02 GB used,  29253.99 GFLOPS
step 2 peak global mem: 128.62777                                                                                                                     
    3  499.22 ms run,  278.50 ms python,  197.64 ms HIP * 6,  23.08 ms fetch data,  7.06 loss, 0.00 acc, 0.001247 LR, 138.02 GB used,  31000.13 GFLOPS
step 3 peak global mem: 128.62777                                                                                                                     
```

So restoring from checkpoint consumes 0.58GB more ram than fresh init in the first step in this case.

And finally, without patch, resume from ckpt, don't realize assign in load_state_dict:

```
RESUME=./ckpts/20240219_121235_bfn1jsys_e37.safe WANDB_RESUME=bfn1jsys BEAM=2 CKPT_EPOCHS=100 BS=624 GPUS=6 MODEL=resnet WORKERS=32 python examples/mlperf/model_train.py 1000
    0 39770.09 ms run, 39466.85 ms python,  280.04 ms HIP * 6,  23.19 ms fetch data,  1.90 loss, 0.78 acc, 0.078467 LR, 2.55 GB used,    389.14 GFLOPS
step 0 peak global mem: 102.26374                                                                                                                     
    1 6321.49 ms run, 6284.05 ms python,    6.87 ms HIP * 6,  30.57 ms fetch data,  1.95 loss, 0.77 acc, 0.078456 LR, 138.02 GB used,   2448.15 GFLOPS
step 1 peak global mem: 128.53910                                                                                                                     
    2  531.86 ms run,  308.80 ms python,  172.13 ms HIP * 6,  50.93 ms fetch data,  1.86 loss, 0.77 acc, 0.078445 LR, 138.02 GB used,  29097.69 GFLOPS
step 2 peak global mem: 128.62658                                                                                                                     
    3  496.33 ms run,  297.61 ms python,  173.38 ms HIP * 6,  25.34 ms fetch data,  1.86 loss, 0.78 acc, 0.078434 LR, 138.02 GB used,  31180.77 GFLOPS
step 3 peak global mem: 128.62659                                                                                                                     
```

Looks almost exactly the same as fresh init.

Now the problem I ran into actually happened after the second step, which looks the same in all of these cases. That crash looked like #3364 #3343. So maybe that is something else, maybe this was just triggered by different behavior below the LRUAllocator level.