@chaosagent 2024-03-09
time linearizer is gated by a cache. So if beam results and timings are already cached, the allocation doesn't happen.